// src/app/models/notes.model.ts
export class Note {
    labelname: string;
    title: String;
    description: String;
    changeColor: String;
    pin: number;
    istrash: number;
    archive: number;
    reminder: string;
    collaborator: string;
    search: number;
    profilePic: string;
    id: number;
    email: string;
}